return types, 64-bit changes: ioctl cmd -> u_long, casts, etc.

+++ sys/kern/subr_log.c
@@ -4,7 +4,7 @@
  *
  * %sccs.include.redist.c%
  *
- *	@(#)subr_log.c	8.1 (Berkeley) 06/10/93
+ *	@(#)subr_log.c	8.2 (Berkeley) 01/09/95
  */
 
 /*
@@ -161,7 +161,7 @@ logwakeup()
 /*ARGSUSED*/
 logioctl(dev, com, data, flag, p)
 	dev_t dev;
-	int com;
+	u_long com;
 	caddr_t data;
 	int flag;
 	struct proc *p;
