+++ sys/kern/tty_pty.c
@@ -4,7 +4,7 @@
  *
  * %sccs.include.redist.c%
  *
- *	@(#)tty_pty.c	8.2 (Berkeley) 09/23/93
+ *	@(#)tty_pty.c	8.3 (Berkeley) 01/09/95
  */
 
 /*
@@ -510,7 +510,7 @@ block:
 /*ARGSUSED*/
 ptyioctl(dev, cmd, data, flag, p)
 	dev_t dev;
-	int cmd;
+	u_long cmd;
 	caddr_t data;
 	int flag;
 	struct proc *p;
